refactoring content analytics screen
