+++ app/(dashboard)/_components/content-analytics/index.tsx
@@ -1,1 +1,1 @@
-export * from './screens/ContentAnalyticsScreen';
+export * from './components/ContentAnalyticsScreen';
