+++ app/(dashboard)/content/page.tsx
@@ -1,4 +1,4 @@
-import { ContentAnalyticsScreen } from '../_components/content-analytics-screen'
+import { ContentAnalyticsScreen } from '../_components/content-analytics/components/ContentAnalyticsScreen'
 
 export default function ContentPage() {
   return <ContentAnalyticsScreen />
